expose network and environment variable (#323)
* expose network and environment variable

+++ builder/builder.go
@@ -39,6 +39,10 @@ func NewBuilder(pm *procmanager.ProcManager, debug bool, workspaceDir string) *B
 // RunTask executes a Task.
 func (b *Builder) RunTask(ctx context.Context, task *graph.Task) error {
 	for _, network := range task.Networks {
+		if network.SkipCreation {
+			log.Printf("Skip creating network: %s\n", network.Name)
+			continue
+		}
 		log.Printf("Creating Docker network: %s, driver: '%s'\n", network.Name, network.Driver)
 		if msg, err := network.Create(ctx, b.procManager); err != nil {
 			return fmt.Errorf("Failed to create network: %s, err: %v, msg: %s", network.Name, err, msg)
@@ -132,6 +136,10 @@ func (b *Builder) CleanTask(ctx context.Context, task *graph.Task) {
 	}
 
 	for _, network := range task.Networks {
+		if network.SkipCreation {
+			log.Printf("Skip deleting network: %s\n", network.Name)
+			continue
+		}
 		if msg, err := network.Delete(ctx, b.procManager); err != nil {
 			log.Printf("Failed to delete network: %s, err: %v, msg: %s\n", network.Name, err, msg)
 		}
